hash_table: move to context-adapter approach
We effectively now follow the internal implementation used by
hash maps in Zig's standard library. With this change, the
contexts are now only concerned with operating on internal keys,
or on a mixture of both, however, it is no longer necessary to
enforce providing method for key insertion by the consumer.