Implement HashSet for use in string tables
The HashSet acts as nothing more but a reverse lookup table.
Here, instead of precomputing `size` to `capacity` mapping for
the HashSet how LLVM is doing, we are cleverly relying on Zig's
`comptime` feature to find the next required capacity value for
a given size programatically (and in compilation time!).

Take this time to also refactor and move around types, e.g., move
hashing functions into `hash.zig` module as they are used by both
the HashTable and the HashSet.